--- frontend/src/components/member/PopUpMSG.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function PopUpMSG() {
-  return <div>PopUpMSG</div>;
-}
-
-export default PopUpMSG;
