--- frontend/src/components/member/PopUpQNA.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-
-function PopUpQNA() {
-  return (
-    <div
-      className="popp-up-qna"
-      style={{ width: "50vw", height: "50vh", background: "white" }}
-    >
-      <h2>QNA</h2>
-      <div className="QNA"></div>
-    </div>
-  );
-}
-
-export default PopUpQNA;
