+++ frontend/src/utils/MemberUtils.js
@@ -1,0 +1,4 @@
+export const formatDate = (date) => {
+  if (!date) return "";
+  return `${date.getFullYear()}-${date.getMonth() + 1}-${date.getDate()}`;
+};
